App Laoyout done!
Hi Guys The App Layout is done! Hence! You have to have all the files in the same map just like it is now (py files, UI files, and source files together) If you separate the files into other maps, the app is not going to work =). Feel free to test it and code =)